Create subseq.cpp
Given a sequence of integers, design an algorithm to generate all of its subsequences. 

 

Note  :  We define a subsequence of a given sequence as a sequence which can be obtained by deleting zero or more elements from the given sequence such that the relative order between the sequence elements doesn't change.