Create arrq2.cpp
Take as input N, the size of array. Take N more inputs and store that in an array. Take as input "target,"  a number. Write a function which prints all pairs of numbers which sum to target.

Input Format
The first line contains input N. Next N lines contains the elements of array and (N+1)th line contains target number.

ength of the arrays should be between 1 and 1000.
Sample Input
5
1
3
4
2
5
5

Sample Output
1 and 4
2 and 3